+++ CHANGELOG.md
@@ -19,6 +19,10 @@ incremented for features.
 
 * lang: Allows one to use `remaining_accounts` with `CpiContext` by implementing the `ToAccountMetas` trait on `CpiContext` ([#351](https://github.com/project-serum/anchor/pull/351/files)).
 
+### Breaking
+
+* lang, ts: Framework defined error codes are introduced, reserving error codes 0-300 for Anchor, and 200-300 for user defined error codes ([#354](https://github.com/project-serum/anchor/pull/354)).
+
 ## [0.7.0] - 2021-05-31
 
 ### Features